+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
 ----------
 - Add class TechnicalAudioMetadata for creating technical audio metadata in AudioMD standard
 - Add class TechnicalVideoMetadata for creating technical video metadata in VideoMD standard
+- Add class TechnicalCSVMetadata for creating technical CSV metadata in ADDML standard
 
 All notable changes to this project will be documented in this file.